Tidy up CMake scripts and allow for out-of-source
Neatify formatting of CMake statements to (generally) follow modern
style, with 2.6 compatibility. Replace CMAKE_SOURCE_DIR with
PROJECT_SOURCE_DIR so that VDT can be embedded in other projects.

Move signature generation section from lib/ to src/ CMakeLists.txt,
implementing it in terms of a custom command and outputting signature
source file to base of binary directory. Ensure that Python interpreter
is located so that generation script can be run! TODO: Do the same
for the VC wrapper, but I don't have VC yet, whatever that is.